Succeeded to change to TableView on FoodViewController and to return the price.
* Change the data-type of productID etc from int to NSInteger. Because to get the value from NSMutableDictionary when Product object is assigned, I couldn’t convert from NSTaggedPointString to int. I don’t know why.

+++ ShoppingCart/Cloth.h
@@ -14,7 +14,7 @@
 
 @property NSArray<Material*> *clothMaterials;
 
-- (id) initWithClothID:(int)clothID
+- (id) initWithClothID:(NSInteger)clothID
              clothName:(NSString*)clothName
             clothPrice:(float)clothPrice
     clothMadeInCountry:(NSString*)clothMadeInCountry
