+++ ShoppingCart/Drink.m
@@ -10,12 +10,12 @@
 
 @implementation Drink
 
-- (id) initWithDrinkID:(int)drinkID
+- (id) initWithDrinkID:(NSInteger)drinkID
              drinkName:(NSString*)drinkName
             drinkPrice:(float)drinkPrice
     drinkMadeInCountry:(NSString*)drinkMadeInCountry
            isDrinkDiet:(BOOL)isDrinkDiet
-             drinkSize:(int)drinkSize {
+             drinkSize:(NSInteger)drinkSize {
     
     self = [super initWithProductID:drinkID productName:drinkName productPrice:drinkPrice productMadeInCountry:drinkMadeInCountry];
