+++ ShoppingCart/Food.m
@@ -10,7 +10,7 @@
 
 @implementation Food
 
-- (id) initWithFoodID:(int)foodID foodName:(NSString*)foodName foodPrice:(float)foodPrice foodMadeInCountry:(NSString*)foodMadeInCountry foodCalorie:(int)foodCalorie foodSize:(int)foodSize foodIngredients:(NSArray<NSString*>*)foodIngredients {
+- (id) initWithFoodID:(NSInteger)foodID foodName:(NSString*)foodName foodPrice:(float)foodPrice foodMadeInCountry:(NSString*)foodMadeInCountry foodCalorie:(NSInteger)foodCalorie foodSize:(NSInteger)foodSize foodIngredients:(NSArray<NSString*>*)foodIngredients {
     
     self = [super initWithProductID:foodID productName:foodName productPrice:foodPrice productMadeInCountry:foodMadeInCountry];
